sync_collections() raises an IOError if Medusa does not return HTTP 200

+++ app/medusa/medusa_ingester.rb
@@ -77,29 +77,36 @@ class MedusaIngester
         'downloading collection list')
     client = MedusaClient.new
     response = client.get(url)
-    struct = JSON.parse(response.body)
 
-    ActiveRecord::Base.transaction do
-      # Create or update a DLS counterpart of each collection.
-      struct.each_with_index do |st, index|
-        col = Collection.find_or_create_by(repository_id: st['uuid'])
-        col.update_from_medusa
+    if response.status == 200
+      struct = JSON.parse(response.body)
+      @@logger.debug("MedusaIngester.sync_collections(): "\
+      "found #{struct.length} collections")
 
-        if task and index % 10 == 0
-          task.percent_complete = index / struct.length.to_f
-          task.save
+      ActiveRecord::Base.transaction do
+        # Create or update a DLS counterpart of each collection.
+        struct.each_with_index do |st, index|
+          col = Collection.find_or_create_by(repository_id: st['uuid'])
+          col.update_from_medusa
+
+          if task and index % 10 == 0
+            task.percent_complete = index / struct.length.to_f
+            task.save
+          end
+        end
+
+        # Delete any DLS collections that are no longer present in Medusa (but
+        # not any items within them, to be safe).
+        Collection.all.each do |col|
+          if struct.select { |st| st['uuid'] == col.repository_id }.empty?
+            @@logger.info('MedusaIngester.sync_collections(): '\
+                "deleting #{col.title} (#{col.repository_id})")
+            col.destroy!
+          end
         end
       end
-
-      # Delete any DLS collections that are no longer present in Medusa (but
-      # not any items within them, to be safe).
-      Collection.all.each do |col|
-        if struct.select { |st| st['uuid'] == col.repository_id }.empty?
-          @@logger.info('MedusaIngester.sync_collections(): '\
-              "deleting #{col.title} (#{col.repository_id})")
-          col.destroy!
-        end
-      end
+    else
+      raise IOError, "GET #{url} returned HTTP #{response.status}"
     end
   end
